+++ src/App.jsx
@@ -1,26 +1,17 @@
 import { Show, createEffect, onMount } from "solid-js";
-import { Route, Routes } from "@solidjs/router";
-import Navbar from "./components/Navbar";
-import Playground from "./components/Playground";
-import Footer from "./components/Footer";
 import { error } from "./stores/error";
-import isUUID from 'validator/es/lib/isUUID'
-import { room } from "./stores/room";
 import { lazy } from "solid-js";
-import { useI18n } from "@solid-primitives/i18n";
+import { room } from "./stores/room";
 import { ui } from "./stores/ui";
 import { settings } from "./stores/settings";
+import Router from "./router";
+import Navbar from "./components/Navbar";
+import Footer from "./components/Footer";
 
-const HomePage = lazy(() => import('./components/HomePage/HomePage'))
 const ErrorModal = lazy(() => import('./components/Modal/ErrorModal'))
 const SettingsModal = lazy(() => import('./components/Modal/SettingsModal'))
 
-const playgroundRouteFilter = {
-  id: (id) => isUUID(id,4)
-}
-
 export default function App() {
-  const [t] = useI18n()
   onMount(() => {
     document.documentElement.setAttribute("lang", settings.lang)
     document.documentElement.className = settings.theme
@@ -34,21 +25,7 @@ export default function App() {
         <Navbar />
       </header>
       <main class={`[ flex-column ] [ align-items-center flex-grow-2 ${settings.theme} ]`}>
-        <Routes>
-          <Route
-            path="/"
-            element={HomePage}
-          />
-          <Route
-            path="/:id"
-            element={Playground}
-            matchFilters={playgroundRouteFilter}
-          />
-          <Route
-            path="*"
-            element={<p class="text-align-center mblock-auto">{t('error.pageNotFound')}</p>}
-          />
-        </Routes>
+        <Router />
       </main>
       <Show when={room.status !== 1}>
         <Footer />
